Refactoring: Group line message creators together and extract a abstract base class from them sot that the interfaces of the creators are consistent.
