+++ src/shared/line/carousel_column_creators.py
@@ -42,38 +42,3 @@ class UqProductManagementTemplateColumnCreator:
         if self._product.is_on_sale:
             return f"原價：NT$ {self._product.original_price}\n特價：NT$ {self._product.special_offer}"
         return f"NT$ {self._product.original_price}"
-
-
-class UqProductManagementTemplateFactory:
-    def get_uq_product_management_template_column_creator(self, *args, **kwargs):
-        return UqProductManagementTemplateColumnCreator(*args, **kwargs)
-
-
-class UqProductManagementTemplateMessageCreator:
-    """
-    Class for creating the carousel template message for user product management.
-    """
-
-    _uq_products: list[UqProduct]
-
-    def __init__(
-        self,
-        products: list[UqProduct],
-        factory: UqProductManagementTemplateFactory = None,
-    ):
-        # Decided not to use DI here because we heavily rely on data models.
-        self._uq_products = products
-        self._factory = (
-            factory if factory is not None else UqProductManagementTemplateFactory()
-        )
-
-    def generate_products_carousel_template_message(self):
-        carousel_template_columns = []
-        for product in self._uq_products:
-            column_creator = (
-                self._factory.get_uq_product_management_template_column_creator(product)
-            )
-            column = column_creator.generate_product_carousel_column()
-            carousel_template_columns.append(column)
-        template = models.CarouselTemplate(columns=carousel_template_columns)
-        return models.TemplateSendMessage(template=template)
